databricks_medicaid_migration/requirements.txt
# Databricks Medicaid Claims Analysis - Requirements
# Python 3.9+

# Core Databricks
databricks-cli>=0.18.0
databricks-connect>=13.0.0
databricks-feature-store>=0.10.0
databricks-sdk>=0.20.0

# Apache Spark (provided by Databricks Runtime)
# pyspark==3.4.0  # Included in DBR

# Delta Lake (provided by Databricks Runtime)
# delta-spark==2.4.0  # Included in DBR

# MLflow
mlflow>=2.10.0
mlflow-skinny>=2.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Data Validation
great-expectations>=0.17.0
pydantic>=2.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# HTTP/API
requests>=2.31.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
pytz>=2023.3

# Cloud Storage
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
adlfs>=2023.10.0  # Azure Data Lake Storage

# Data Quality
pandas-profiling>=3.6.0
datacompy>=0.10.0

# Testing (for development)
pytest>=7.4.0
pytest-spark>=0.6.0
chispa>=0.9.0  # Spark DataFrame testing

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
pylint>=3.0.0
mypy>=1.8.0

# Jupyter (for local development)
jupyter>=1.0.0
ipython>=8.18.0
notebook>=7.0.0
